requirements.txt
PyJWT
Requests
fastapi
uvicorn
pytest
python-dotenv
python-jose
pytest-cov